feature (single sector commission)
- prints the receipt of a single sector commissions fetched from the database
